pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "engauto-mcp"
version = "0.1.0"
description = "Engineering Automation MCP Server"
requires-python = ">=3.10"
dependencies = [
  "aiosqlite>=0.20.0",
  "mcp>=1.0.0",
  "pydantic>=2.0.0",
  "python-json-patch>=1.33",
]

[project.scripts]
engauto-mcp-server = "engauto_mcp.__main__:main"
engauto-mcp-client = "engauto_mcp.client:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
